Add tests for measurements
