+++ test.js
@@ -1,2 +1,3 @@
 import './tests/nodes.test.js'
 import './tests/transactions.test.js'
+import './tests/measure.test.js'
